Move infer_registry into a more public spot
